nettoyage

+++ app/src/main/java/org/oucho/musicplayer/utils/ColorUtils.java
@@ -4,7 +4,7 @@ package org.oucho.musicplayer.utils;
 public class ColorUtils {
 
     public static int applyAlpha(int color, float alpha) {
-        int colorAlpha  = (int)((color & 0xFF000000) >> 24) + 256;
+        int colorAlpha  = ((color & 0xFF000000) >> 24) + 256;
         return (color & 0x00FFFFFF) | ((int) (colorAlpha * alpha) << 24);
     }
 }
